高并发 秒杀 超卖问题的解决方案： redis+批量sql语句持久化方案，使用junit单元测试，通过多线程来模拟并发
程序入口：UserRedPacketServiceTest、RedisApplication、application-seckill.yml

流程
1、抢红包服务userRedPacketService
2、访问redis抢红包，具体来说
  	a. 访问库存，大于0时执行b
  	b. 更新库存，向redis插入抢红包记录，
  这两步操作要有原子性，从而保证可重复读。
3、持久化
  	a. 访问redis获取抢红包记录
  	b. 将记录转换成DTO，再通过Mybatis插入到数据库，并更新库存

实践过程中遇到的问题：
1、junit下，资源的GC、掉线问题：因为我是使用junit、多线程模拟并发的。必须要保证主线程在所有的线程结束之后再结束。不然就会出现各种timeout、shutdown、closed。原因不明，大概和主线程结束、spring容器中的资源被回收有关？（不过如果走controller，走web，应该就不会遇到这种问题了，datasource应该不会再自动关闭，自动回收？）
      大概遇到过这些错误：
      a. pool提前关闭，导致持久化未能完成，Mysql事务回滚。出现这种情况的时候，有时有任何的报错，让人很莫名奇妙。
      INFO 14852 --- [ionShutdownHook] com.zaxxer.hikari.HikariDataSource       : HikariPool-1 - Shutdown initiated...
      可能会提示这些错误吧：
          - java.sql.SQLException: No operations allowed after statement closed.
          - Could not get a resource from the pool; nested exception is io.lettuce.core

2、调试中，因为timeout设置不够，也会在debug的过程中出现问题，如redis:timeout:

3、redisTemplate和底层api的序列化问题。显然获取byte[]的底层api不会自动进行序列化。
通过Lettuce使用Lua脚本的注意事项：
    Lua脚本不会进行再次编码，传入什么就写入什么。所以必须要用redisTemplate的转码器转码，再传入Lua脚本。
    不然我们后面再通过redisTemplate从redis取数据的时候，是会进行反序列化的，这就会导致编码对不上，从而各种nil，各种报错，例如：
    - Could not read JSON: Unexpected character ('-' (code 44)): Expected space separating root-level values

4、启动了太多的其他的组件。可以通过配置yml解决——spring:autoconfigure:exclude:

参考：https://blog.csdn.net/zzqtty/article/details/81901325
 * 高并发抢红包案列以及使用锁，版本号，redis缓存解决，项目可运行，详细注释(三)

+++ demo-base/src/main/java/com/demo/base/example/redis/service/UserRedPacketService.java
@@ -8,5 +8,5 @@ public interface UserRedPacketService {
      * @param userId 用户编号
      * @return
      */
-    int grabRedPacket(Integer redPacketId, Integer userId);
+    Long grabRedPacket(Integer redPacketId, Integer userId);
 }
